+++ CHANGELOG.md
@@ -11,6 +11,18 @@ the following form: YYYY.0M.0D.
 
 ### Added
 
+- Implemented kernel class `ArtificialNeuron1L` for the ArtificialNeuron
+  component
+- Selected `ArrayList<Double>` as the representation
+- Defined convention and correspondence for the representation
+- Implemented all kernel methods (`addWeight`, `setWeight`, `getWeight`,
+  `weightCount`) and Standard methods (`clear`, `newInstance`,
+  `transferFrom`)
+
+## 2026.04.24
+
+### Added
+
 - Designed abstract class `ArtificialNeuronSecondary` for the
   ArtificialNeuron component
 - Implemented secondary methods (`scaleWeights`, `sumWeights`,